+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilterTableModel.java
@@ -228,8 +228,8 @@ final class DefaultFilterTableModel<R, C> extends AbstractTableModel implements 
 		}
 
 		@Override
-		public String string(int rowIndex, C identifier) {
-			return columns.string(items.included().get(rowIndex), requireNonNull(identifier));
+		public String format(int rowIndex, C identifier) {
+			return columns.format(items.included().get(rowIndex), requireNonNull(identifier));
 		}
 
 		@Override
@@ -317,9 +317,9 @@ final class DefaultFilterTableModel<R, C> extends AbstractTableModel implements 
 
 		private boolean accepts(R item, ConditionModel<?> condition, C identifier, TableColumns<R, C> columns) {
 			if (condition.valueClass().equals(String.class)) {
-				String string = columns.string(item, identifier);
+				String formatted = columns.format(item, identifier);
 
-				return ((ConditionModel<String>) condition).accepts(string.isEmpty() ? null : string);
+				return ((ConditionModel<String>) condition).accepts(formatted.isEmpty() ? null : formatted);
 			}
 
 			return condition.accepts(columns.comparable(item, identifier));